fix(backend, users, auth): fixed errors with auth refreshing tokens, and validatin error in user register

+++ backend/src/auth/service.py
@@ -1,7 +1,7 @@
 import uuid
 from datetime import UTC, datetime, timedelta
 
-from src.auth.exceptions import TokenExpired, UserAuthDenied
+from src.auth.exceptions import CookieError, TokenExpired, UserAuthDenied
 from src.auth.repository import AuthRepository
 from src.auth.schemas import RefreshSessionCreate, Tokens
 from src.auth.utils import generate_jwt_token
@@ -69,8 +69,12 @@ class AuthService:
             access_token=access_token, refresh_token=refresh_session.refresh_token
         )
 
-    async def refresh_token(self, refresh_token: uuid.UUID) -> Tokens:
-        session = await self.repository.get_refresh_session(refresh_token)
+    async def refresh_token(self, refresh_token: str | None) -> Tokens:
+        if refresh_token is None:
+            raise CookieError("Refresh token is empty")
+
+        token = uuid.UUID(hex=refresh_token)
+        session = await self.repository.get_refresh_session(token)
 
         await self.repository.delete(session)
